# Improve memory and cpu performance