- add a simple check on the sha to abort re-marshaling the index if the sha hasnt changed.
- adds some simple benchmarks to validate this change.

pre-change
```
go test -bench=. -v -benchmem github.com/fluxcd/source-controller/internal/helm/repository -run=^a -benchtime=10s
goos: linux
goarch: amd64
pkg: github.com/fluxcd/source-controller/internal/helm/repository
cpu: AMD FX(tm)-6100 Six-Core Processor
BenchmarkHelmChartIndex
BenchmarkHelmChartIndex-6   	   15674	    765526 ns/op	   47414 B/op	     883 allocs/op
PASS
ok  	github.com/fluxcd/source-controller/internal/helm/repository	19.777s
```

post-change
```
go test -bench=. -v -benchmem github.com/fluxcd/source-controller/internal/helm/repository -run=^a -benchtime=10s
goos: linux
goarch: amd64
pkg: github.com/fluxcd/source-controller/internal/helm/repository
cpu: AMD FX(tm)-6100 Six-Core Processor
BenchmarkHelmChartIndex
BenchmarkHelmChartIndex-6   	  621212	     18005 ns/op	     304 B/op	       6 allocs/op
PASS
ok  	github.com/fluxcd/source-controller/internal/helm/repository	11.451s
```