@tehlers320 2022-03-19
Could i get some guidance from the maintainers on how to best approach this?  I'm not sure if there is some history the existing cache system. The performance issue is in unmarshaling large indexes over and over again (ours is 30mb and will grow). 

I threw in a cache for fun and everything is fixed in a big way now. I'm just not sure how to proceed. 
https://github.com/tehlers320/source-controller/commit/9b62349427fb457525c2587580047e77943a06b7

Performance is fantastic now 

With in memory cache
```
Reconciliation finished in 737.117902ms, next run in 10m0s
```

Without in memory cache (avoiding marshal)
```
Reconciliation finished in 7.910032747s, next run in 10m0s 
```

Here is what throwing a memory cache in looks like, CPU is cut dramatically. Memory not at all. Test build put in at 21:25 (was v0.21.2 prior)
<img width="458" alt="Screen Shot 2022-03-18 at 9 43 08 PM" src="https://user-images.githubusercontent.com/16598119/159103776-fbe01078-5bfd-4364-8541-05ce4192512d.png">
<img width="468" alt="Screen Shot 2022-03-18 at 9 43 30 PM" src="https://user-images.githubusercontent.com/16598119/159103784-07d85e7f-24d5-4c3c-a93a-240a085c9216.png">