@souleb 2022-03-19
@tehlers320 I am also testing adding a cache to the controller that would cache an arbitrary number of indexes with expiration. Let's talk about it on slack.