@tehlers320 2022-03-20
Closing in favor of https://github.com/fluxcd/source-controller/pull/626/files